adding call redirection

+++ src/main/java/com/websathi/connectmeapp/model/business/Review.java
@@ -17,5 +17,5 @@ public class Review {
     public String id;
     public String name;
     public String comment;
-    public Integer rating;
+    public Double rating;
 }
